requirements.txt
accelerate==1.7.0
aiohttp==3.12.13
backports.tarfile==1.2.0
datasets==3.6.0
decord==0.6.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
pip-chill==1.0.3
qwen-omni-utils==0.0.8
tomli==2.0.1
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
transformers==4.52.4
